Address bug in issue #117
The original fix for #117 had a slight bug that was affecting the create
Point GRG by Reference System. If the user tried to create a point GRG
within a polar region with a Grid Size that was not Grid Zone the GRG
would not be created and no warning was given.

On investigation if the grid origin point was in a polar region the
mgrs.USNGtoPoint() function returns a point that has NaN values for its
x and y. Added a check to exit the function with a warning message if
this occurs.

Also used this PR to address a slight issue with the order of the extent
graphics layer prevent the user from adjusting it because it was
displayed under the GRGarea layer.